Moved helper functions to individual files
